[refactor] Merge users-email-change tests into users.test.ts
